[chore] #105 Change sub folder in feature's performance module
