+++ feature/performance/src/main/java/com/cmc/curtaincall/feature/performance/lostitem/LostItemTypeGrid.kt
@@ -1,4 +1,4 @@
-package com.cmc.curtaincall.feature.performance.ui.lostitem
+package com.cmc.curtaincall.feature.performance.lostitem
 
 import androidx.annotation.DrawableRes
 import androidx.compose.foundation.background
