+++ feature/performance/src/main/java/com/cmc/curtaincall/feature/performance/lostitem/PerformanceLostItemCreateScreen.kt
@@ -1,4 +1,4 @@
-package com.cmc.curtaincall.feature.performance.ui.lostitem
+package com.cmc.curtaincall.feature.performance.lostitem
 
 import androidx.compose.foundation.BorderStroke
 import androidx.compose.foundation.background
