+++ feature/performance/src/main/java/com/cmc/curtaincall/feature/performance/lostitem/PerformanceLostItemScreen.kt
@@ -1,4 +1,4 @@
-package com.cmc.curtaincall.feature.performance.ui.lostitem
+package com.cmc.curtaincall.feature.performance.lostitem
 
 import androidx.compose.foundation.*
 import androidx.compose.foundation.layout.*
