+++ feature/performance/src/main/java/com/cmc/curtaincall/feature/performance/lostitem/PerformanceLostItemTabScreen.kt
@@ -1,6 +1,7 @@
-package com.cmc.curtaincall.feature.performance.ui.review
+package com.cmc.curtaincall.feature.performance.lostitem
 
 import androidx.compose.foundation.BorderStroke
+import androidx.compose.foundation.Image
 import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
@@ -9,23 +10,23 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import com.cmc.curtaincall.common.design.R
-import com.cmc.curtaincall.common.design.component.custom.RatingBar
 import com.cmc.curtaincall.common.design.extensions.toSp
 import com.cmc.curtaincall.common.design.theme.*
 
 @Composable
-internal fun PerformanceReviewTabScreen(
+internal fun PerformanceLostItemTabScreen(
     modifier: Modifier = Modifier,
-    onNavigateReview: () -> Unit
+    onNavigateLostItem: () -> Unit
 ) {
     Column(modifier.padding(horizontal = 20.dp)) {
-        Row {
+        Row(verticalAlignment = Alignment.CenterVertically) {
             Text(
-                text = "총 324개의 한 줄 리뷰",
+                text = stringResource(R.string.performance_find_lost_item),
                 color = Chinese_Black,
                 fontSize = 17.dp.toSp(),
                 fontWeight = FontWeight.Bold,
@@ -36,7 +37,7 @@ internal fun PerformanceReviewTabScreen(
                 modifier = Modifier
                     .size(71.dp, 28.dp)
                     .border(BorderStroke(1.dp, Me_Pink), RoundedCornerShape(14.dp))
-                    .clickable { onNavigateReview() },
+                    .clickable { onNavigateLostItem() },
                 contentAlignment = Alignment.Center
             ) {
                 Text(
@@ -48,14 +49,14 @@ internal fun PerformanceReviewTabScreen(
                 )
             }
         }
-        Column(Modifier.padding(top = 14.dp, bottom = 20.dp)) {
-            PerformanceReviewSimpleItem(Modifier.fillMaxWidth())
-            PerformanceReviewSimpleItem(
+        Column(Modifier.padding(top = 14.dp, bottom = 59.dp)) {
+            PerformanceLostItem(Modifier.fillMaxWidth())
+            PerformanceLostItem(
                 modifier = Modifier
                     .fillMaxWidth()
                     .padding(top = 12.dp)
             )
-            PerformanceReviewSimpleItem(
+            PerformanceLostItem(
                 modifier = Modifier
                     .fillMaxWidth()
                     .padding(top = 12.dp)
@@ -65,34 +66,47 @@ internal fun PerformanceReviewTabScreen(
 }
 
 @Composable
-private fun PerformanceReviewSimpleItem(
+private fun PerformanceLostItem(
     modifier: Modifier = Modifier
 ) {
-    Column(
+    Row(
         modifier = modifier
             .border(BorderStroke(1.dp, Bright_Gray), RoundedCornerShape(10.dp))
-            .padding(start = 12.dp, top = 14.dp, bottom = 16.dp)
+            .padding(10.dp),
+        verticalAlignment = Alignment.CenterVertically
     ) {
-        RatingBar(
-            modifier = Modifier.size(14.dp),
-            rating = 4
+        Image(
+            painter = painterResource(R.drawable.img_default_performance),
+            contentDescription = null,
+            modifier = Modifier.size(63.dp)
         )
-        Text(
-            text = "zkvpzkvpzk | 2023.6.28",
-            modifier = Modifier.padding(top = 6.dp),
-            color = Black_Coral,
-            fontSize = 13.dp.toSp(),
-            fontWeight = FontWeight.Medium,
-            fontFamily = spoqahansanseeo
-        )
-        Text(
-            text = "최고령 리어왕'으로 기네스북에 오를 예정이라시는~!",
-            modifier = Modifier.padding(top = 12.dp),
-            color = Nero,
-            fontSize = 13.dp.toSp(),
-            fontWeight = FontWeight.Medium,
-            fontFamily = spoqahansanseeo,
-            lineHeight = 16.dp.toSp()
-        )
+        Column(Modifier.padding(start = 12.dp)) {
+            Row {
+                Text(
+                    text = "닥스 지갑",
+                    color = Chinese_Black,
+                    fontSize = 14.dp.toSp(),
+                    fontWeight = FontWeight.Bold,
+                    fontFamily = spoqahansanseeo
+                )
+                Spacer(Modifier.weight(1f))
+                Text(
+                    text = "2023.6.15",
+                    color = Roman_Silver,
+                    fontSize = 10.dp.toSp(),
+                    fontWeight = FontWeight.Normal,
+                    fontFamily = spoqahansanseeo
+                )
+            }
+            Spacer(Modifier.weight(1f))
+            Text(
+                text = "습득 장소 | LG아트센터 서울\n보관 장소 | LG아트센터 서울",
+                color = Black_Coral,
+                fontSize = 12.dp.toSp(),
+                fontWeight = FontWeight.Medium,
+                fontFamily = spoqahansanseeo,
+                lineHeight = 16.dp.toSp()
+            )
+        }
     }
 }
