+++ feature/performance/src/main/java/com/cmc/curtaincall/feature/performance/review/PerformanceReviewScreen.kt
@@ -1,4 +1,4 @@
-package com.cmc.curtaincall.feature.performance.ui.review
+package com.cmc.curtaincall.feature.performance.review
 
 import androidx.compose.foundation.*
 import androidx.compose.foundation.layout.*
